use trait, docs blocks, make subscriptions cache lifetime configurable & more
