+++ app/config/parameters.yml.dist
@@ -53,3 +53,5 @@ parameters:
     env(PAYWALL_SERVER_URL):
     env(PAYWALL_SERVER_USERNAME):
     env(PAYWALL_SERVER_PASSWORD):
+
+    env(SUBSCRIPTIONS_CACHE_LIFETIME): 86400 # 24 hours
