+++ src/SWP/Bundle/CoreBundle/Model/Article.php
@@ -19,10 +19,11 @@ namespace SWP\Bundle\CoreBundle\Model;
 use SWP\Bundle\ContentBundle\Model\Article as BaseArticle;
 use SWP\Component\MultiTenancy\Model\OrganizationAwareTrait;
 use SWP\Component\MultiTenancy\Model\TenantAwareTrait;
+use SWP\Component\Paywall\Model\PaywallSecuredTrait;
 
 class Article extends BaseArticle implements ArticleInterface
 {
-    use TenantAwareTrait, OrganizationAwareTrait;
+    use TenantAwareTrait, OrganizationAwareTrait, PaywallSecuredTrait;
 
     /**
      * @var PackageInterface
@@ -43,11 +44,6 @@ class Article extends BaseArticle implements ArticleInterface
      * @var ExternalArticleInterface
      */
     protected $externalArticle;
-
-    /**
-     * @var bool
-     */
-    protected $paywallSecured = false;
 
     /**
      * {@inheritdoc}
@@ -120,15 +116,5 @@ class Article extends BaseArticle implements ArticleInterface
     public function setExternalArticle(ExternalArticleInterface $externalArticle): void
     {
         $this->externalArticle = $externalArticle;
-    }
-
-    public function isPaywallSecured(): bool
-    {
-        return $this->paywallSecured;
-    }
-
-    public function setPaywallSecured(bool $paywallSecured): void
-    {
-        $this->paywallSecured = $paywallSecured;
     }
 }
