+++ src/SWP/Bundle/CoreBundle/Model/User.php
@@ -16,19 +16,12 @@ declare(strict_types=1);
 
 namespace SWP\Bundle\CoreBundle\Model;
 
-use Doctrine\Common\Collections\Collection;
 use SWP\Bundle\UserBundle\Model\User as BaseUser;
 use SWP\Component\MultiTenancy\Model\OrganizationAwareTrait;
-use SWP\Component\Paywall\Model\SubscriptionInterface;
 
 class User extends BaseUser implements UserInterface
 {
     use OrganizationAwareTrait;
-
-    /**
-     * @var Collection
-     */
-    protected $subscriptions;
 
     /**
      * {@inheritdoc}
@@ -65,34 +58,5 @@ class User extends BaseUser implements UserInterface
     public function getSubscriberId(): string
     {
         return $this->email;
-    }
-
-    public function getSubscriptions(): Collection
-    {
-        return $this->subscriptions;
-    }
-
-    public function setSubscriptions(Collection $subscriptions): void
-    {
-        $this->subscriptions = $subscriptions;
-    }
-
-    public function hasSubscription(SubscriptionInterface $subscription): bool
-    {
-        return $this->subscriptions->contains($subscription);
-    }
-
-    public function addSubscription(SubscriptionInterface $subscription): void
-    {
-        if (!$this->hasSubscription($subscription)) {
-            $this->subscriptions->add($subscription);
-        }
-    }
-
-    public function removeSubscription(SubscriptionInterface $subscription): void
-    {
-        if ($this->hasSubscription($subscription)) {
-            $this->subscriptions->removeElement($subscription);
-        }
     }
 }
