+++ src/SWP/Bundle/CoreBundle/Provider/CachedSubscriptionsProvider.php
@@ -5,11 +5,17 @@ declare(strict_types=1);
 namespace SWP\Bundle\CoreBundle\Provider;
 
 use Doctrine\Common\Cache\Cache;
+use SWP\Component\MultiTenancy\Context\TenantContextInterface;
 use SWP\Component\Paywall\Model\SubscriberInterface;
 
 final class CachedSubscriptionsProvider implements SubscriptionsProviderInterface
 {
     public const CACHE_KEY_PREFIX = 'subscriptions_provider_';
+
+    /**
+     * @var int
+     */
+    private $cacheLifeTime;
 
     /**
      * @var SubscriptionsProviderInterface
@@ -21,10 +27,21 @@ final class CachedSubscriptionsProvider implements SubscriptionsProviderInterfac
      */
     private $cacheProvider;
 
-    public function __construct(SubscriptionsProviderInterface $decoratedProvider, Cache $cacheProvider)
-    {
+    /**
+     * @var TenantContextInterface
+     */
+    private $tenantContext;
+
+    public function __construct(
+        int $cacheLifeTime,
+        SubscriptionsProviderInterface $decoratedProvider,
+        Cache $cacheProvider,
+        TenantContextInterface $tenantContext
+    ) {
+        $this->cacheLifeTime = $cacheLifeTime;
         $this->decoratedProvider = $decoratedProvider;
         $this->cacheProvider = $cacheProvider;
+        $this->tenantContext = $tenantContext;
     }
 
     public function getSubscriptions(SubscriberInterface $subscriber, array $filters = []): array
@@ -36,7 +53,7 @@ final class CachedSubscriptionsProvider implements SubscriptionsProviderInterfac
         } else {
             $subscriptions = $this->decoratedProvider->getSubscriptions($subscriber, $filters);
 
-            $this->cacheProvider->save($cacheKey, $subscriptions, 86400);
+            $this->cacheProvider->save($cacheKey, $subscriptions, $this->cacheLifeTime);
         }
 
         return $subscriptions;
@@ -44,6 +61,6 @@ final class CachedSubscriptionsProvider implements SubscriptionsProviderInterfac
 
     private function generateCacheKey(SubscriberInterface $subscriber): string
     {
-        return self::CACHE_KEY_PREFIX.$subscriber->getSubscriberId();
+        return self::CACHE_KEY_PREFIX.$this->tenantContext->getTenant()->getCode().$subscriber->getSubscriberId();
     }
 }
