--- src/SWP/Bundle/CoreBundle/Resources/config/doctrine-orm/Subscription.orm.yml
@@ -1,10 +1,0 @@
-SWP\Bundle\CoreBundle\Model\Subscription:
-    type: mappedSuperclass
-    table: swp_subscription
-    manyToOne:
-        user:
-            targetEntity: SWP\Bundle\CoreBundle\Model\UserInterface
-            joinColumn:
-                name: user_id
-                referencedColumnName: id
-                nullable: false
