+++ src/SWP/Bundle/CoreBundle/Resources/config/doctrine-orm/User.orm.yml
@@ -8,7 +8,3 @@ SWP\Bundle\CoreBundle\Model\User:
                 name: organization_id
                 referencedColumnName: id
                 nullable: false
-    oneToMany:
-        subscriptions:
-            targetEntity: SWP\Component\Paywall\Model\SubscriptionInterface
-            mappedBy: user
