+++ src/SWP/Bundle/PaywallBundle/DependencyInjection/Configuration.php
@@ -16,12 +16,7 @@ declare(strict_types=1);
 
 namespace SWP\Bundle\PaywallBundle\DependencyInjection;
 
-use SWP\Bundle\PaywallBundle\Doctrine\ORM\SubscriptionRepository;
-use SWP\Bundle\StorageBundle\Doctrine\ORM\EntityRepository;
 use SWP\Component\Paywall\Adapter\PaymentsHubAdapter;
-use SWP\Component\Paywall\Model\Subscription;
-use SWP\Component\Paywall\Model\SubscriptionInterface;
-use SWP\Component\Storage\Factory\Factory;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
@@ -38,32 +33,6 @@ class Configuration implements ConfigurationInterface
                 ->scalarNode('adapter')
                     ->defaultValue(PaymentsHubAdapter::class)
                     ->info('Subscriptions System Adapter')
-                ->end()
-                ->arrayNode('persistence')
-                    ->addDefaultsIfNotSet()
-                    ->children()
-                        ->arrayNode('orm')
-                            ->addDefaultsIfNotSet()
-                            ->canBeEnabled()
-                            ->children()
-                                ->arrayNode('classes')
-                                    ->addDefaultsIfNotSet()
-                                    ->children()
-                                        ->arrayNode('subscription')
-                                            ->addDefaultsIfNotSet()
-                                            ->children()
-                                                ->scalarNode('model')->cannotBeEmpty()->defaultValue(Subscription::class)->end()
-                                                ->scalarNode('repository')->defaultValue(SubscriptionRepository::class)->end()
-                                                ->scalarNode('interface')->defaultValue(SubscriptionInterface::class)->end()
-                                                ->scalarNode('factory')->defaultValue(Factory::class)->end()
-                                                ->scalarNode('object_manager_name')->defaultValue(null)->end()
-                                            ->end()
-                                        ->end()
-                                    ->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                    ->end()
                 ->end()
             ->end();
