--- src/SWP/Bundle/PaywallBundle/Resources/config/doctrine-orm/Subscription.orm.yml
@@ -1,36 +1,0 @@
-SWP\Component\Paywall\Model\Subscription:
-    type: mappedSuperclass
-    table: swp_subscription
-    gedmo:
-        soft_deleteable:
-            field_name: deletedAt
-            time_aware: false
-        loggable: true
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-        code:
-            type: string
-            unique: true
-        type:
-            type: string
-        details:
-            type: array
-        createdAt:
-            type: datetime
-            gedmo:
-                timestampable:
-                    on: create
-        updatedAt:
-            type: datetime
-            nullable: true
-            gedmo:
-                timestampable:
-                    on: update
-        deletedAt:
-            type: datetime
-            nullable: true
